Update Assessment3.sol Location Variable
Realised we should have Location,

Added it as another input to register alongside with in the DidRegistry.

Also created validation checks in the AssignRecipientsbyDid to make sure the location inputted in that function is the same as the addresses already assigned with a location from the DidRegistry. The tokenID struct is then updated with the location.